# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/payment_terms:
    get:
      tags:
        - PaymentTerms
      summary: List PaymentTerms
      description: Get a paginated list of PaymentTerms with optional filtering and sorting
      operationId: listPaymentTerms
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentTermList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PaymentTerms
      summary: Create PaymentTerm
      description: Create a new PaymentTerm
      operationId: createPaymentTerm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentTermRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentTerm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/payment_term_lines:
    get:
      tags:
        - PaymentTermLines
      summary: List PaymentTermLines
      description: Get a paginated list of PaymentTermLines with optional filtering and sorting
      operationId: listPaymentTermLines
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentTermLineList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PaymentTermLines
      summary: Create PaymentTermLine
      description: Create a new PaymentTermLine
      operationId: createPaymentTermLine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentTermLineRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentTermLine'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PaymentTerm:
      type: object
      required:
        - id
        - name
        - sequence
        - status
        - early_discount
        - discount_percent
        - discount_days
        - discount_tax_basis
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        note:
          nullable: true
          type: string
          maxLength: 500
        sequence:
          type: integer
          format: int32
        status:
          type: string
        early_discount:
          type: boolean
        discount_percent:
          type: string
        discount_days:
          type: integer
          format: int32
        discount_account_id:
          nullable: true
          type: string
          format: uuid
        discount_tax_basis:
          type: string
        metadata:
          type: object
...
    CreatePaymentTermRequest:
      type: object
      required:
        - name
        - sequence
        - status
        - early_discount
        - discount_percent
        - discount_days
        - discount_tax_basis
        - metadata
      properties:
        company_id:
          nullable: true
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        note:
          nullable: true
          type: string
          maxLength: 500
        sequence:
          type: integer
          format: int32
        status:
          type: string
        early_discount:
          type: boolean
        discount_percent:
          type: string
        discount_days:
          type: integer
          format: int32
        discount_account_id:
          nullable: true
          type: string
          format: uuid
        discount_tax_basis:
          type: string
        metadata:
          type: object
...
    PaymentTermList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentTerm'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaymentTermLine:
      type: object
      required:
        - id
        - term_id
        - value
        - value_amount
        - nb_days
        - delay_type
        - anchor
        - sequence
        - metadata
      properties:
        id:
          type: string
          format: uuid
        term_id:
          type: string
          format: uuid
        company_id:
          nullable: true
          type: string
          format: uuid
        value:
          type: string
        value_amount:
          type: string
        nb_days:
          type: integer
          format: int32
        day_of_month:
          nullable: true
          type: integer
          format: int32
        delay_type:
          type: string
        anchor:
          type: string
        sequence:
          type: integer
          format: int32
        metadata:
          type: object
...
    CreatePaymentTermLineRequest:
      type: object
      required:
        - term_id
        - value
        - value_amount
        - nb_days
        - delay_type
        - anchor
        - sequence
        - metadata
      properties:
        term_id:
          type: string
          format: uuid
        company_id:
          nullable: true
          type: string
          format: uuid
        value:
          type: string
        value_amount:
          type: string
        nb_days:
          type: integer
          format: int32
        day_of_month:
          nullable: true
          type: integer
          format: int32
        delay_type:
          type: string
        anchor:
          type: string
        sequence:
          type: integer
          format: int32
        metadata:
          type: object
...
    PaymentTermLineList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PaymentTermLine'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []